Move dryRun to Node
so that we can have nodes for transforms.
